+++ src/grandorgue/dialogs/settings/GOSettingsOrgans.h
@@ -10,12 +10,17 @@
 
 #include <wx/panel.h>
 
+#include "ptrvector.h"
+
 class GOMidi;
-class GOOrgan;
-class GOConfig;
 class wxButton;
 class wxListEvent;
 class wxListView;
+class wxTextCtrl;
+
+class GOArchiveFile;
+class GOConfig;
+class GOOrgan;
 
 class GOSettingsOrgans : public wxPanel {
   enum {
@@ -24,26 +29,43 @@ class GOSettingsOrgans : public wxPanel {
     ID_ORGAN_DOWN,
     ID_ORGAN_UP,
     ID_ORGAN_TOP,
-    ID_ORGAN_PROPERTIES,
-    ID_PACKAGES,
-    ID_PACKAGE_DEL,
+    ID_ORGAN_PROPERTIES
   };
 
 private:
+  struct OrganRec {
+    const GOOrgan *p_organ;
+    bool is_selected;
+    bool is_focused;
+  };
+  using OrganRecs = std::vector<OrganRec>;
+
   GOConfig &m_config;
   GOMidi &m_midi;
+  ptr_vector<GOOrgan> &m_OrigOrganList;
+  std::vector<GOArchiveFile *> m_PackageList;
 
   wxListView *m_Organs;
+  wxTextCtrl *m_Builder;
+  wxTextCtrl *m_Recording;
+  wxTextCtrl *m_OrganHash;
+  wxTextCtrl *m_PackageId;
+  wxTextCtrl *m_PackageName;
+  wxTextCtrl *m_PathInPackage;
+  wxTextCtrl *m_PackageHash;
+  wxTextCtrl *m_PackageInfo;
+
   wxButton *m_OrganDown;
   wxButton *m_OrganUp;
   wxButton *m_OrganTop;
   wxButton *m_OrganDel;
   wxButton *m_OrganProperties;
-  wxListView *m_Packages;
-  wxButton *m_PackageDel;
 
+  OrganRecs GetCurrentOrganRecs();
+  void ReorderOrgans(const OrganRecs &newSortedRecs);
   void MoveOrgan(long from, long to);
 
+  void OnOrganFocused(wxListEvent &event);
   void OnOrganSelected(wxListEvent &event);
   void OnOrganUp(wxCommandEvent &event);
   void OnOrganDown(wxCommandEvent &event);
@@ -51,12 +73,10 @@ private:
   void OnOrganDel(wxCommandEvent &event);
   void OnOrganProperties(wxCommandEvent &event);
 
-  void OnPackageSelected(wxListEvent &event);
-  void OnPackageDel(wxCommandEvent &event);
-
 public:
   GOSettingsOrgans(GOConfig &settings, GOMidi &midi, wxWindow *parent);
 
+  virtual bool TransferDataToWindow() override;
   virtual bool TransferDataFromWindow() override;
 
   DECLARE_EVENT_TABLE()